+++ src/integrationTest/java/uk/gov/hmcts/reform/pip/data/management/service/schemavalidation/nonstrategic/FttLandRegistryWeeklyHearingListTest.java
@@ -29,13 +29,13 @@ import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
 @ActiveProfiles("integration-basic")
 @SpringBootTest
-class FftLandRegistryWeeklyHearingListTest extends IntegrationBasicTestBase {
+class FttLandRegistryWeeklyHearingListTest extends IntegrationBasicTestBase {
 
     @Autowired
     ValidationService validationService;
 
-    private static final String VALID_JSON = "data/non-strategic/fft-land-registry-tribunal-weekly-hearing-list/"
-        + "fftLandRegistryTribunalWeeklyHearingList.json";
+    private static final String VALID_JSON = "data/non-strategic/ftt-land-registry-tribunal-weekly-hearing-list/"
+        + "fttLandRegistryTribunalWeeklyHearingList.json";
     private static final String INVALID_MESSAGE = "Invalid JSON list marked as valid";
 
     private static final String SOURCE_ARTEFACT_ID = "sourceArtefactId";
@@ -58,7 +58,7 @@ class FftLandRegistryWeeklyHearingListTest extends IntegrationBasicTestBase {
     @BeforeEach
     void setup() {
         headerGroup = new HeaderGroup(PROVENANCE, SOURCE_ARTEFACT_ID, ARTEFACT_TYPE, SENSITIVITY, LANGUAGE,
-                                      DISPLAY_FROM, DISPLAY_TO, ListType.FFT_LR_WEEKLY_HEARING_LIST, COURT_ID,
+                                      DISPLAY_FROM, DISPLAY_TO, ListType.FTT_LR_WEEKLY_HEARING_LIST, COURT_ID,
                                       CONTENT_DATE);
     }
