+++ src/main/java/uk/gov/hmcts/reform/pip/data/management/service/ListConversionFactory.java
@@ -58,8 +58,8 @@ import static uk.gov.hmcts.reform.pip.model.publication.ListType.CST_WEEKLY_HEAR
 import static uk.gov.hmcts.reform.pip.model.publication.ListType.ET_DAILY_LIST;
 import static uk.gov.hmcts.reform.pip.model.publication.ListType.ET_FORTNIGHTLY_PRESS_LIST;
 import static uk.gov.hmcts.reform.pip.model.publication.ListType.FAMILY_DAILY_CAUSE_LIST;
-import static uk.gov.hmcts.reform.pip.model.publication.ListType.FFT_LR_WEEKLY_HEARING_LIST;
-import static uk.gov.hmcts.reform.pip.model.publication.ListType.FFT_TAX_WEEKLY_HEARING_LIST;
+import static uk.gov.hmcts.reform.pip.model.publication.ListType.FTT_LR_WEEKLY_HEARING_LIST;
+import static uk.gov.hmcts.reform.pip.model.publication.ListType.FTT_TAX_WEEKLY_HEARING_LIST;
 import static uk.gov.hmcts.reform.pip.model.publication.ListType.GRC_WEEKLY_HEARING_LIST;
 import static uk.gov.hmcts.reform.pip.model.publication.ListType.IAC_DAILY_LIST;
 import static uk.gov.hmcts.reform.pip.model.publication.ListType.IAC_DAILY_LIST_ADDITIONAL_CASES;
@@ -170,13 +170,13 @@ public class ListConversionFactory {
             new NonStrategicListFileConverter(),
             new NonStrategicListSummaryData(PAAC_WEEKLY_HEARING_LIST)
         )),
-        Map.entry(FFT_TAX_WEEKLY_HEARING_LIST, new ConversionPair(
+        Map.entry(FTT_TAX_WEEKLY_HEARING_LIST, new ConversionPair(
             new NonStrategicListFileConverter(),
-            new NonStrategicListSummaryData(FFT_TAX_WEEKLY_HEARING_LIST)
+            new NonStrategicListSummaryData(FTT_TAX_WEEKLY_HEARING_LIST)
         )),
-        Map.entry(FFT_LR_WEEKLY_HEARING_LIST, new ConversionPair(
+        Map.entry(FTT_LR_WEEKLY_HEARING_LIST, new ConversionPair(
             new NonStrategicListFileConverter(),
-            new NonStrategicListSummaryData(FFT_LR_WEEKLY_HEARING_LIST)
+            new NonStrategicListSummaryData(FTT_LR_WEEKLY_HEARING_LIST)
         )),
         Map.entry(RPT_EASTERN_WEEKLY_HEARING_LIST, new ConversionPair(
             new NonStrategicListFileConverter(),
