+++ src/main/resources/application.yaml
@@ -145,8 +145,8 @@ validations:
     SIAC_WEEKLY_HEARING_LIST: "schemas/non-strategic/siac_weekly_hearing_list.json"
     POAC_WEEKLY_HEARING_LIST: "schemas/non-strategic/siac_weekly_hearing_list.json"
     PAAC_WEEKLY_HEARING_LIST: "schemas/non-strategic/siac_weekly_hearing_list.json"
-    FFT_TAX_WEEKLY_HEARING_LIST: "schemas/non-strategic/fft-tax-chamber-tribunal-weekly-hearing-list.json"
-    FFT_LR_WEEKLY_HEARING_LIST: "schemas/non-strategic/fft-land-registry-tribunal-weekly-hearing-list.json"
+    FTT_TAX_WEEKLY_HEARING_LIST: "schemas/non-strategic/ftt-tax-chamber-tribunal-weekly-hearing-list.json"
+    FTT_LR_WEEKLY_HEARING_LIST: "schemas/non-strategic/ftt-land-registry-tribunal-weekly-hearing-list.json"
     RPT_EASTERN_WEEKLY_HEARING_LIST: "schemas/non-strategic/fft-residential-property-tribunal-weekly-hearing-list.json"
     RPT_LONDON_WEEKLY_HEARING_LIST: "schemas/non-strategic/fft-residential-property-tribunal-weekly-hearing-list.json"
     RPT_MIDLANDS_WEEKLY_HEARING_LIST: "schemas/non-strategic/fft-residential-property-tribunal-weekly-hearing-list.json"
