+++ src/test/java/uk/gov/hmcts/reform/pip/data/management/service/filegeneration/nonstrategic/FttTaxWeeklyHearingListConverterTest.java
@@ -20,11 +20,11 @@ import java.nio.charset.Charset;
 import java.util.Map;
 import java.util.Objects;
 
-import static uk.gov.hmcts.reform.pip.model.publication.ListType.FFT_LR_WEEKLY_HEARING_LIST;
+import static uk.gov.hmcts.reform.pip.model.publication.ListType.FTT_TAX_WEEKLY_HEARING_LIST;
 
 @ActiveProfiles("test")
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class FftLandRegistryWeeklyHearingListFileConverterTest {
+class FttTaxWeeklyHearingListConverterTest {
 
     private static final String CONTENT_DATE = "12 December 2024";
     private static final String PROVENANCE = "provenance";
@@ -36,9 +36,9 @@ class FftLandRegistryWeeklyHearingListFileConverterTest {
     private static final String ENGLISH = "ENGLISH";
     private static final String WELSH = "WELSH";
 
-    private static final String LIST_NAME = FFT_LR_WEEKLY_HEARING_LIST.name();
-    private static final String LIST_ENGLISH_NAME = "First-tier Tribunal "
-        + "(Lands Registration Tribunal) Weekly Hearing List";
+    private static final String LIST_NAME = FTT_TAX_WEEKLY_HEARING_LIST.name();
+    private static final String LIST_ENGLISH_NAME = "First-tier Tribunal (Tax Chamber)"
+        + " Weekly Hearing List";
     private static final String LIST_DATE_ENGLISH = "List for 12 December 2024";
     private static final String LIST_DATE_WELSH = "Rhestr ar gyfer 12 December 2024";
     private static final String OBSERVE_HEARING_ENGLISH = "Observe a court or tribunal hearing as a journalist, "
@@ -46,11 +46,14 @@ class FftLandRegistryWeeklyHearingListFileConverterTest {
     private static final String OBSERVE_HEARING_WELSH = "Arsylwi gwrandawiad llys neu dribiwnlys fel newyddiadurwr, "
         + "ymchwilydd neu aelod o'r cyhoedd";
     private static final String LIST_WELSH_NAME = "Rhestr o Wrandawiadau "
-        + "Wythnosol Tribiwnlys Haen Gyntaf (Tribiwnlys Cofrestru Tir)";
+        + "Wythnosol Tribiwnlys Haen Gyntaf (Siambr Treth)";
 
     private static final String HEADER_ELEMENT = "page-heading";
     private static final String LIST_DATE_ELEMENT = "list-date";
-    private static final String CONTACT_MESSAGE_ELEMENT = "contact-message";
+    private static final String CONTACT_MESSAGE_ELEMENT_1 = "contact-message-1";
+    private static final String CONTACT_MESSAGE_ELEMENT_2 = "contact-message-2";
+    private static final String CONTACT_MESSAGE_ELEMENT_3 = "contact-message-3";
+    private static final String CONTACT_MESSAGE_ELEMENT_4 = "contact-message-4";
     private static final String OBSERVE_HEARING_ELEMENT =  "observe-hearing";
 
     private static final String TITLE_MESSAGE = "Title does not match";
@@ -66,18 +69,18 @@ class FftLandRegistryWeeklyHearingListFileConverterTest {
     @BeforeAll
     void setup() throws IOException {
         try (InputStream inputStream = getClass()
-            .getResourceAsStream("/mocks/non-strategic/fftLandRegistryTribunalWeeklyHearingList.json")) {
+            .getResourceAsStream("/mocks/non-strategic/fttTaxWeeklyHearingList.json")) {
             String inputRaw = IOUtils.toString(inputStream, Charset.defaultCharset());
             listInputJson = new ObjectMapper().readTree(inputRaw);
         }
     }
 
     @Test
-    void testLandRegistryWeeklyHearingListFileConversionInEnglish() throws IOException {
+    void testTaxWeeklyHearingListFileConversionInEnglish() throws IOException {
         Map<String, Object> languageResource;
         try (InputStream languageFile = Thread.currentThread()
             .getContextClassLoader()
-            .getResourceAsStream("templates/languages/en/non-strategic/fftLrWeeklyHearingList.json")) {
+            .getResourceAsStream("templates/languages/en/non-strategic/fttTaxWeeklyHearingList.json")) {
             languageResource = new ObjectMapper().readValue(
                 Objects.requireNonNull(languageFile).readAllBytes(), new TypeReference<>() {
                 });
@@ -108,16 +111,31 @@ class FftLandRegistryWeeklyHearingListFileConverterTest {
             .extracting(Element::text)
             .isEqualTo(LIST_DATE_ENGLISH);
 
-        softly.assertThat(document.getElementById(CONTACT_MESSAGE_ELEMENT))
+        softly.assertThat(document.getElementById(CONTACT_MESSAGE_ELEMENT_1))
             .as(BODY_MESSAGE)
             .extracting(Element::text)
-            .isEqualTo("Members of the public wishing to observe a hearing "
-                           + "or representatives of the media may, on their request, join any "
-                           + "telephone or video hearing remotely while they are taking place by sending an "
-                           + "email in advance to the tribunal at [insert office email] with the following details "
-                           + "in the subject line “[OBSERVER/MEDIA] REQUEST – [case reference] – [hearing date] "
-                           + "(need to include any other infromation required by the tribunal)” and appropriate "
-                           + "arrangements will be made to allow access where reasonably practicable");
+            .isEqualTo("Open justice is a fundamental principle of our justice system. You can attend a "
+                           + "public hearing in person, or you can apply for permission to observe remotely.");
+        softly.assertThat(document.getElementById(CONTACT_MESSAGE_ELEMENT_2))
+            .as(BODY_MESSAGE)
+            .extracting(Element::text)
+            .isEqualTo("Members of the public and the media can ask to join any telephone or "
+                           + "video hearing remotely. Contact the Tribunal before the hearing to ask for "
+                           + "permission to attend by emailing taxappeals@justice.gov.uk.");
+
+        softly.assertThat(document.getElementById(CONTACT_MESSAGE_ELEMENT_3))
+            .as(BODY_MESSAGE)
+            .extracting(Element::text)
+            .isEqualTo("The subject line for the email should contain the following wording: "
+                           + "“HEARING ACCESS REQUEST – [Appellant’s name] v [Respondent’s name, for example HMRC] – "
+                           + "[case reference] – [hearing date]”. You will be sent instructions on "
+                           + "how to join the hearing.");
+
+        softly.assertThat(document.getElementById(CONTACT_MESSAGE_ELEMENT_4))
+            .as(BODY_MESSAGE)
+            .extracting(Element::text)
+            .isEqualTo("The judge may refuse a request and can also decide a "
+                           + "hearing must be held in private, in such cases you will not be able to attend.");
 
         softly.assertThat(document.getElementById(OBSERVE_HEARING_ELEMENT))
             .as(BODY_MESSAGE)
@@ -126,14 +144,15 @@ class FftLandRegistryWeeklyHearingListFileConverterTest {
 
         softly.assertThat(document.getElementsByTag("th"))
             .as(TABLE_HEADERS_MESSAGE)
-            .hasSize(6)
+            .hasSize(7)
             .extracting(Element::text)
             .containsExactly(
                 "Date",
                 "Hearing time",
                 "Case name",
                 "Case reference number",
-                "Judge",
+                "Judge(s)",
+                "Member(s)",
                 "Venue/Platform"
             );
 
@@ -141,11 +160,11 @@ class FftLandRegistryWeeklyHearingListFileConverterTest {
     }
 
     @Test
-    void testLandRegistryWeeklyHearingListFileConversionInWelsh() throws IOException {
+    void testTaxWeeklyHearingListFileConversionInWelsh() throws IOException {
         Map<String, Object> languageResource;
         try (InputStream languageFile = Thread.currentThread()
             .getContextClassLoader()
-            .getResourceAsStream("templates/languages/cy/non-strategic/fftLrWeeklyHearingList.json")) {
+            .getResourceAsStream("templates/languages/cy/non-strategic/fttTaxWeeklyHearingList.json")) {
             languageResource = new ObjectMapper().readValue(
                 Objects.requireNonNull(languageFile).readAllBytes(), new TypeReference<>() {
                 });
@@ -176,17 +195,33 @@ class FftLandRegistryWeeklyHearingListFileConverterTest {
             .extracting(Element::text)
             .isEqualTo(LIST_DATE_WELSH);
 
-        softly.assertThat(document.getElementById(CONTACT_MESSAGE_ELEMENT))
+        softly.assertThat(document.getElementById(CONTACT_MESSAGE_ELEMENT_1))
             .as(BODY_MESSAGE)
             .extracting(Element::text)
-            .isEqualTo("Gall aelodau o’r cyhoedd sy’n dymuno arsylwi gwrandawiad neu "
-                           + "gynrychiolwyr y cyfryngau ymuno ag unrhyw wrandawiad dros y ffôn neu "
-                           + "drwy fideo o bell ar gais tra’u bod yn cael eu cynnal drwy anfon "
-                           + "e-bost ymlaen llaw at y tribiwnlys yn [insert office email] gyda’r manylion "
-                           + "canlynol yn y llinell bwnc “CAIS [ARSYLLWR/CYFRYNGAU] – [cyfeirnod yr achos] – "
-                           + "[dyddiad y gwrandawiad] (angen cynnwys unrhyw wybodaeth arall sy’n ofynnol "
-                           + "gan y tribiwnlys)” a gwneir trefniadau priodol i ganiatáu mynediad "
-                           + "lle bo hynny’n rhesymol ymarferol.");
+            .isEqualTo("Mae cyfiawnder agored yn egwyddor sylfaenol yn ein system gyfiawnder. "
+                           + "Gallwch fynychu gwrandawiad cyhoeddus wyneb yn wyneb, neu gallwch "
+                           + "wneud cais am ganiatâd i arsylwi o bell.");
+        softly.assertThat(document.getElementById(CONTACT_MESSAGE_ELEMENT_2))
+            .as(BODY_MESSAGE)
+            .extracting(Element::text)
+            .isEqualTo("Gall aelodau’r cyhoedd a’r cyfryngau ofyn am gael ymuno ag "
+                           + "unrhyw wrandawiad dros y ffôn neu drwy fideo o bell. Cysylltwch â’r "
+                           + "Tribiwnlys cyn y gwrandawiad i ofyn am ganiatâd i fod yn bresennol "
+                           + "drwy anfon e-bost at taxappeals@justice.gov.uk");
+
+        softly.assertThat(document.getElementById(CONTACT_MESSAGE_ELEMENT_3))
+            .as(BODY_MESSAGE)
+            .extracting(Element::text)
+            .isEqualTo("Dylai llinell bwnc yr e-bost gynnwys y geiriad canlynol: "
+                           + "“CAIS MYNEDIAD I WRANDAWIAD – [Enw’r apelydd] v [Enw’r atebydd, er enghraifft CThEF] – "
+                           + "[cyfeirnod yr achos] – [dyddiad y gwrandawiad]”. Anfonir cyfarwyddiadau atoch ar "
+                           + "sut i ymuno â’r gwrandawiad.");
+
+        softly.assertThat(document.getElementById(CONTACT_MESSAGE_ELEMENT_4))
+            .as(BODY_MESSAGE)
+            .extracting(Element::text)
+            .isEqualTo("Gall y barnwr wrthod cais a gall hefyd benderfynu bod yn rhaid cynnal "
+                           + "gwrandawiad yn breifat, ac mewn achosion o'r fath ni fyddwch yn gallu bod yn bresennol.");
 
         softly.assertThat(document.getElementById(OBSERVE_HEARING_ELEMENT))
             .as(BODY_MESSAGE)
@@ -195,14 +230,15 @@ class FftLandRegistryWeeklyHearingListFileConverterTest {
 
         softly.assertThat(document.getElementsByTag("th"))
             .as(TABLE_HEADERS_MESSAGE)
-            .hasSize(6)
+            .hasSize(7)
             .extracting(Element::text)
             .containsExactly(
                 "Dyddiad",
                 "Amser y gwrandawiad",
                 "Enw’r achos",
                 "Cyfeirnod yr achos",
-                "Barnwr",
+                "Barnwr/Barnwyr",
+                "Aelod(au)",
                 "Lleoliad/Platfform"
             );
