+++ src/test/resources/mocks/non-strategic/fttTaxWeeklyHearingList.json
